@@ -1,7 +1,7 @@
 {
   "name": "cate-desktop",
   "productName": "Cate Desktop",
-  "version": "1.0.1-dev.1",
+  "version": "1.1.0-dev.1",
   "description": "A graphical user interface for the ESA CCI Toolbox (Cate)",
   "main": "main.js",
   "author": "CCI Toolbox (Cate) Development Team",
@@ -15,15 +15,18 @@
     "classnames": "^2.2.5",
     "d3": "^3.5.17",
     "deep-equal": "^1.0.1",
+    "deep-object-diff": "^1.0.4",
     "electron-devtools-installer": "^2.2.1",
     "oboe": "^2.1.3",
-    "openlayers": "^4.2.0",
+    "openlayers": "^4.6.4",
     "proj4": "^2.4.4",
-    "react": "^15.6.1",
-    "react-addons-css-transition-group": "^15.6.0",
-    "react-dom": "^15.6.1",
-    "react-redux": "^5.0.5",
-    "redux": "^3.6.0",
+    "react": "^16.2.0",
+    "react-color": "^2.13.8",
+    "react-dom": "^16.2.0",
+    "react-redux": "^5.0.6",
+    "react-transition-group": "^1.2.1",
+    "react-addons-css-transition-group": "^15.6.2",
+    "redux": "^3.7.2",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.2.0",
     "reselect": "^2.5.4",
