@@ -1,7 +1,7 @@
 {
   "name": "cate-desktop",
   "productName": "Cate Desktop",
-  "version": "1.0.1-dev.1",
+  "version": "1.1.0-dev.1",
   "description": "A graphical user interface for the ESA CCI Toolbox (Cate)",
   "private": true,
   "scripts": {
@@ -53,35 +53,38 @@
     "@types/oboe": "^2.0.28",
     "@types/openlayers": "^3.20.6",
     "@types/proj4": "^2.3.4",
-    "@types/react": "^15.0.36",
-    "@types/react-dom": "^15.5.1",
-    "@types/react-redux": "^4.4.46",
-    "@types/react-test-renderer": "^15.5.1",
-    "@types/redux-logger": "^3.0.0",
+    "@types/react": "^16.0.28",
+    "@types/react-dom": "^16.0.3",
+    "@types/react-redux": "^5.0.14",
+    "@types/react-test-renderer": "^16.0.0",
+    "@types/redux-logger": "^3.0.5",
     "@types/reselect": "^2.0.27",
     "@types/tether": "^1.1.27",
     "cesium": "^1.40.0",
-    "chai": "^3.5.0",
-    "chai-as-promised": "^6.0.0",
+    "cesium-typings": "^1.40.1",
+    "chai": "^4.1.2",
+    "chai-as-promised": "^7.1.1",
     "chai-enzyme": "^0.8.0",
     "classnames": "^2.2.5",
     "d3": "^3.5.17",
     "deep-equal": "^1.0.1",
+    "deep-object-diff": "^1.0.4",
     "electron": "1.7.9",
     "electron-builder": "19.48.3",
     "electron-devtools-installer": "^2.2.1",
-    "enzyme": "^2.6.0",
-    "mocha": "^3.2.0",
-    "oboe": "^2.1.3",
-    "openlayers": "^4.2.0",
+    "enzyme": "^3.2.0",
+    "mocha": "^4.0.1",
+    "oboe": "^2.1.4",
+    "openlayers": "^4.6.4",
     "proj4": "^2.4.4",
-    "react": "^15.6.1",
-    "react-addons-css-transition-group": "^15.6.0",
-    "react-addons-perf": "^15.4.2",
-    "react-dom": "^15.6.1",
-    "react-redux": "^5.0.5",
-    "react-test-renderer": "^15.6.1",
-    "redux": "^3.7.1",
+    "react": "^16.2.0",
+    "react-color": "^2.13.8",
+    "react-dom": "^16.2.0",
+    "react-redux": "^5.0.6",
+    "react-transition-group": "^1.2.1",
+    "react-addons-css-transition-group": "^15.6.2",
+    "react-test-renderer": "^16.2.0",
+    "redux": "^3.7.2",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.2.0",
     "reselect": "^2.5.4",
