@@ -18,6 +18,7 @@ export const DATA_FRAME_LIKE_TYPE = 'cate.core.types.DataFrameLike';
 export const POINT_LIKE_TYPE = 'cate.core.types.PointLike';
 export const POLYGON_LIKE_TYPE = 'cate.core.types.PolygonLike';
 export const GEOMETRY_LIKE_TYPE = 'cate.core.types.GeometryLike';
+export const GEO_DATA_FRAME_PROXY_TYPE = 'cate.core.types.GeoDataFrame';
 export const TIME_LIKE_TYPE = 'cate.core.types.TimeLike';
 export const TIME_RANGE_LIKE_TYPE = 'cate.core.types.TimeRangeLike';
 export const VAR_NAME_LIKE_TYPE = 'cate.core.types.VarName';
@@ -26,6 +27,17 @@ export const DICT_LIKE_TYPE = 'cate.core.types.DictLike';
 export const FILE_LIKE_TYPE = 'cate.core.types.FileLike';
 export const ARBITRARY_TYPE = 'cate.core.types.Arbitrary';
 export const LITERAL_TYPE = 'cate.core.types.Literal';
+
+export const GEOMETRY_TYPE = 'shapely.geometry.base.BaseGeometry';
+export const POINT_TYPE = 'shapely.geometry.point.Point';
+export const MULTI_POINT_TYPE = 'shapely.geometry.multipoint.MultiPoint';
+export const LINE_STRING_TYPE = 'shapely.geometry.linestring.LineString';
+export const MULTI_LINE_STRING_TYPE = 'shapely.geometry.multilinestring.MultiLineString';
+export const POLYGON_TYPE = 'shapely.geometry.polygon.Polygon';
+export const MULTI_POLYGON_TYPE = 'shapely.geometry.multipolygon.multiPolygon';
+export const GEOMETRY_COLLECTION_TYPE = 'shapely.geometry.collection.GeometryCollection';
+
+
 
 /**
  * Naive test if a targetDataType is assignable from a given sourceDataType.
@@ -56,9 +68,13 @@ export function isAssignableFrom(targetDataType: string, sourceDataType: string)
         case ND_ARRAY_TYPE:
             return sourceDataType === DATA_ARRAY_TYPE;
         case DATA_FRAME_TYPE:
-            return sourceDataType === DATA_FRAME_LIKE_TYPE || sourceDataType === GEO_DATA_FRAME_TYPE;
+            return sourceDataType === DATA_FRAME_LIKE_TYPE || sourceDataType === GEO_DATA_FRAME_TYPE
+                || sourceDataType === GEO_DATA_FRAME_PROXY_TYPE;
         case DATA_FRAME_LIKE_TYPE:
-            return sourceDataType === DATA_FRAME_TYPE || sourceDataType === GEO_DATA_FRAME_TYPE;
+            return sourceDataType === DATA_FRAME_TYPE || sourceDataType === GEO_DATA_FRAME_TYPE
+                || sourceDataType === GEO_DATA_FRAME_PROXY_TYPE;
+        case GEO_DATA_FRAME_TYPE:
+            return sourceDataType === GEO_DATA_FRAME_PROXY_TYPE;
         case DATASET_TYPE:
             return sourceDataType === DATASET_LIKE_TYPE;
         case DATASET_LIKE_TYPE:
@@ -78,16 +94,21 @@ export function isAssignableFrom(targetDataType: string, sourceDataType: string)
         case TIME_RANGE_LIKE_TYPE:
             return sourceDataType === STR_TYPE;
         case POINT_LIKE_TYPE:
-            // TODO (forman): add shapely Point
-            return sourceDataType === STR_TYPE;
+            return sourceDataType === STR_TYPE || sourceDataType === POINT_TYPE;
         case POLYGON_LIKE_TYPE:
-            // TODO (forman): add shapely Polygon
-            return sourceDataType === STR_TYPE;
+            return sourceDataType === STR_TYPE || sourceDataType === POLYGON_TYPE;
         case GEOMETRY_LIKE_TYPE:
-            // TODO (forman): add shapely Point, Polygon, Geometry, ...
             return sourceDataType === STR_TYPE
                 || sourceDataType === POINT_LIKE_TYPE
-                || sourceDataType === POLYGON_LIKE_TYPE;
+                || sourceDataType === POLYGON_LIKE_TYPE
+                || sourceDataType === GEOMETRY_TYPE
+                || sourceDataType === POINT_TYPE
+                || sourceDataType === MULTI_POINT_TYPE
+                || sourceDataType === LINE_STRING_TYPE
+                || sourceDataType === MULTI_LINE_STRING_TYPE
+                || sourceDataType === POLYGON_TYPE
+                || sourceDataType === MULTI_POLYGON_TYPE
+                || sourceDataType === GEOMETRY_COLLECTION_TYPE;
     }
     return false;
 }
