@@ -4,16 +4,18 @@ import {toTextValue} from "./Field";
 import {TextField, TextFieldValue} from "./TextField";
 import {GeometryDialog} from "../GeometryDialog";
 import {GeometryType, validateGeometryValue} from "../../../common/geometry-util";
+import {GeometryWKTGetter} from "../../containers/editor/ValueEditor";
 
 
 interface IGeometryFieldProps {
-    value: TextFieldValue | any,
+    value: TextFieldValue | any;
     onChange: (value: TextFieldValue) => void;
     geometryType: GeometryType;
     placeholder?: string;
     size?: number;
     nullable?: boolean;
     disabled?: boolean;
+    geometryWKTGetter?: GeometryWKTGetter;
 }
 
 interface IGeometryFieldState {
@@ -26,10 +28,22 @@ export class GeometryField extends React.Component<IGeometryFieldProps, IGeometr
         super(props);
         this.validateGeometryText = this.validateGeometryText.bind(this);
         this.state = {isEditorOpen: false};
+        this.setGeometryWKT = this.setGeometryWKT.bind(this);
     }
 
     validateGeometryText(value: string  | null) {
         validateGeometryValue(value, this.props.geometryType);
+    }
+
+    setGeometryWKT() {
+        let wkt;
+        try {
+            wkt = this.props.geometryWKTGetter && this.props.geometryWKTGetter();
+        } catch (error) {
+            console.error(error);
+            wkt = "Error: " + error;
+        }
+        this.props.onChange({textValue: wkt, value: wkt});
     }
 
     render() {
@@ -53,6 +67,10 @@ export class GeometryField extends React.Component<IGeometryFieldProps, IGeometr
                     validator={this.validateGeometryText}
                     nullable={this.props.nullable}
                 />
+
+                <AnchorButton className="pt-intent-primary pt-icon-selection" style={{flex: 'none'}}
+                              disabled={!this.props.geometryWKTGetter}
+                              onClick={this.setGeometryWKT}/>
 
                 <AnchorButton className="pt-intent-primary" style={{flex: 'none'}}
                               onClick={() => this.setState({isEditorOpen: true})}>...</AnchorButton>
