@@ -13,6 +13,7 @@ import {VarNameValueEditor} from "./VarNameValueEditor";
 import {DictValueEditor} from "./DictValueEditor";
 import {LiteralValueEditor} from "./LiteralValueEditor";
 import {TimeValueEditor} from "./TimeValueEditor";
+import * as Cesium from "cesium";
 
 export interface InputAssignment {
     constantValue: FieldValue<any> | any;
@@ -25,12 +26,15 @@ export type InputAssignments = { [inputName: string]: InputAssignment };
 export type ValueEditorValue<T> = FieldValue<T> | T | null;
 export type ValueEditorCallback<T> = (input: OperationInputState, value: ValueEditorValue<T>) => void;
 
+export type GeometryWKTGetter = (() => string) | null;
+
 export interface IValueEditorProps<T> {
     input: OperationInputState;
     value: ValueEditorValue<T>;
     onChange: ValueEditorCallback<T>;
     inputAssignments?: InputAssignments;
     resources?: ResourceState[];
+    geometryWKTGetter?: GeometryWKTGetter;
 }
 
 export type ValueEditorFactory<T> = (props: IValueEditorProps<T>) => JSX.Element | null;
@@ -95,17 +99,17 @@ function renderTextValueEditor(props: IValueEditorProps<string>) {
 
 function renderPointLikeValueEditor(props: IValueEditorProps<string>) {
     return <GeometryValueEditor input={props.input} value={props.value} onChange={props.onChange} geometryType="Point"
-                                size={16}/>;
+                                size={16} geometryWKTGetter={props.geometryWKTGetter}/>;
 }
 
 function renderPolygonLikeValueEditor(props: IValueEditorProps<string>) {
     return <GeometryValueEditor input={props.input} value={props.value} onChange={props.onChange} geometryType="Polygon"
-                                size={32}/>;
+                                size={32} geometryWKTGetter={props.geometryWKTGetter}/>;
 }
 
 function renderGeometryLikeValueEditor(props: IValueEditorProps<string>) {
     return <GeometryValueEditor input={props.input} value={props.value} onChange={props.onChange}
-                                geometryType="Geometry" size={32}/>;
+                                geometryType="Geometry" size={32} geometryWKTGetter={props.geometryWKTGetter}/>;
 }
 
 function renderTimeLikeValueEditor(props: IValueEditorProps<string>) {
