@@ -1,22 +1,25 @@
+import {combineReducers, Reducer} from 'redux';
+import deepEqual = require("deep-equal");
 import {
-    State, DataState, LocationState, SessionState, CommunicationState, ControlState, DataStoreState,
-    LayerState, Placemark
+State, DataState, LocationState, SessionState, CommunicationState, ControlState, DataStoreState,
+LayerState, Placemark, VectorLayerBase
 } from './state';
 import * as actions from './actions';
+import {Action} from "./actions";
 import * as assert from "../common/assert";
-import {combineReducers, Reducer} from 'redux';
 import {updateObject, updatePropertyObject} from "../common/objutil";
 import {
     SELECTED_VARIABLE_LAYER_ID, updateSelectedVariableLayer,
     newWorldView, newTableView, newFigureView, getFigureViewTitle, genPlacemarkId,
-    hasWebGL,
+    hasWebGL, isVectorLayer, PLACEMARKS_LAYER,
 } from "./state-util";
 import {
-    removeViewFromLayout, removeViewFromViewArray, ViewState, addViewToViewArray,
-    addViewToLayout, selectViewInLayout, getViewPanel, findViewPanel, splitViewPanel, changeViewSplitPos,
-    addViewToPanel, moveView, selectView
+removeViewFromLayout, removeViewFromViewArray, ViewState, addViewToViewArray,
+addViewToLayout, selectViewInLayout, getViewPanel, findViewPanel, splitViewPanel, changeViewSplitPos,
+addViewToPanel, moveView, selectView
 } from "./components/ViewState";
-import {Action, SET_GLOBE_MOUSE_POSITION, SET_GLOBE_VIEW_POSITION} from "./actions";
+import {isString} from "../common/types";
+import {featurePropertiesFromSimpleStyle} from "../common/geojson-simple-style";
 
 // Note: reducers are unit-tested through actions.spec.ts
 
@@ -111,23 +114,10 @@ const dataReducer = (state: DataState = initialDataState, action: Action) => {
 const initialView = newWorldView();
 
 const initialControlState: ControlState = {
-    selectedDataStoreId: null,
-    selectedDataSourceId: null,
-    dataSourceFilterExpr: '',
-    selectedOperationName: null,
-    operationFilterTags: [],
-    operationFilterExpr: '',
     selectedWorkflowStepId: null,
     selectedWorkspaceResourceName: null,
     selectedVariableName: null,
     dialogs: {},
-
-    showDataSourceDetails: true,
-    showResourceDetails: true,
-    showWorkflowStepDetails: true,
-    showOperationDetails: true,
-    showVariableDetails: true,
-    showLayerDetails: true,
 
     views: [initialView],
     viewLayout: {
@@ -137,6 +127,8 @@ const initialControlState: ControlState = {
     activeViewId: initialView.id,
 
     worldViewClickAction: null,
+
+    entityUpdateCount: 0,
 };
 
 
@@ -145,13 +137,26 @@ const controlReducer = (state: ControlState = initialControlState, action: Actio
         case actions.RENAME_RESOURCE: {
             const resName = action.payload.resName;
             const newResName = action.payload.newResName;
+
+            // Rename resName in selectedWorkspaceResourceName
             let selectedWorkspaceResourceName = state.selectedWorkspaceResourceName;
             if (selectedWorkspaceResourceName === resName) {
                 selectedWorkspaceResourceName = newResName;
             }
+
+            // Rename resName in views and their layers
             const views = viewsReducer(state.views, action, state.activeViewId);
-            if (selectedWorkspaceResourceName !== state.selectedWorkspaceResourceName || views !== state.views) {
+
+            if (selectedWorkspaceResourceName !== state.selectedWorkspaceResourceName
+                || views !== state.views) {
                 return {...state, selectedWorkspaceResourceName, views};
+            }
+            return state;
+        }
+        case actions.UPDATE_ENTITY_STYLE: {
+            const views = viewsReducer(state.views, action, state.activeViewId);
+            if (views !== state.views) {
+                return {...state, views};
             }
             return state;
         }
@@ -264,6 +269,10 @@ const controlReducer = (state: ControlState = initialControlState, action: Actio
             const viewLayout = changeViewSplitPos(state.viewLayout, viewPath, delta);
             return {...state, viewLayout};
         }
+        case actions.INC_ENTITY_UPDATE_COUNT: {
+            const entityUpdateCount = state.entityUpdateCount + 1;
+            return {...state, entityUpdateCount};
+        }
         default: {
             const newViews = viewsReducer(state.views, action, state.activeViewId);
             if (newViews !== state.views) {
@@ -329,6 +338,16 @@ const viewReducer = (state: ViewState<any>, action: Action, activeViewId: string
             }
             break;
         }
+        case actions.UPDATE_ENTITY_STYLE: {
+            const viewId = action.payload.viewId;
+            if (state.id === viewId && state.type === 'world') {
+                const layers = layersReducer(state.data.layers, action, isActiveView);
+                if (layers !== state.data.layers) {
+                    return {...state, data: {...state.data, layers}};
+                }
+            }
+            break;
+        }
         case actions.SET_SHOW_SELECTED_VARIABLE_LAYER: {
             if (state.type === 'world') {
                 const layers = layersReducer(state.data.layers, action, isActiveView);
@@ -338,13 +357,27 @@ const viewReducer = (state: ViewState<any>, action: Action, activeViewId: string
             }
             break;
         }
-        // TODO (forman): clean up code duplication here, they are basically all the same
-        //                SET_SELECTED_LAYER_ID, SET_VIEW_MODE, SET_PROJECTION_CODE,SET_SPLIT_LAYER_ID, SET_SPLIT_LAYER_POS
+        // TODO (forman): clean up code duplication here, following actions are basically all the same
+        //                SET_SELECTED_ENTITY_ID,
+        //                SET_SELECTED_LAYER_ID,
+        //                SET_VIEW_MODE,
+        //                SET_PROJECTION_CODE,
+        //                SET_SPLIT_LAYER_ID,
+        //                SET_SPLIT_LAYER_POS
+        //
         case actions.SET_SELECTED_LAYER_ID: {
             const viewId = action.payload.viewId;
             if (viewId === state.id && state.type === 'world') {
                 const selectedLayerId = action.payload.selectedLayerId;
                 return {...state, data: {...state.data, selectedLayerId}};
+            }
+            break;
+        }
+        case actions.SET_SELECTED_ENTITY_ID: {
+            const viewId = action.payload.viewId;
+            if (viewId === state.id && state.type === 'world') {
+                const selectedEntityId = action.payload.selectedEntityId;
+                return {...state, data: {...state.data, selectedEntityId}};
             }
             break;
         }
@@ -461,6 +494,20 @@ const viewReducer = (state: ViewState<any>, action: Action, activeViewId: string
             }
             break;
         }
+        case actions.UPDATE_LAYER_STYLE: {
+            const viewId = action.payload.viewId;
+            if (viewId === state.id) {
+                assert.ok(state.type === 'world');
+                const {layerId, style} = action.payload;
+                const layers = state.data.layers.slice();
+                const layerIndex = layers.findIndex(l => l.id === layerId);
+                assert.ok(layerIndex >= 0, "layerIndex >= 0");
+                let layer = layers[layerIndex];
+                layers[layerIndex] = {...layer, style: {...layer.style, ...style}};
+                return {...state, data: {...state.data, layers}};
+            }
+            break;
+        }
         case actions.UPDATE_TABLE_VIEW_DATA: {
             const viewId = action.payload.viewId;
             if (viewId === state.id) {
@@ -510,6 +557,19 @@ const layerReducer = (state: LayerState, action: Action, isActiveView: boolean) 
             }
             break;
         }
+        case actions.UPDATE_ENTITY_STYLE: {
+            const {layerId, entityId, style} = action.payload;
+            if (state.id === layerId && isVectorLayer(state)) {
+                const vectorLayer = state as VectorLayerBase;
+                const entityStyles = vectorLayer.entityStyles;
+                const oldStyle = entityStyles && entityStyles[entityId];
+                const newStyle = {...oldStyle, ...style};
+                if (!deepEqual(oldStyle, newStyle)) {
+                    return {...state, entityStyles: {...entityStyles, [entityId]: newStyle}};
+                }
+            }
+            break;
+        }
         case actions.SET_SHOW_SELECTED_VARIABLE_LAYER: {
             if (state.id === SELECTED_VARIABLE_LAYER_ID) {
                 const showSelectedVariableLayer = action.payload.showSelectedVariableLayer;
@@ -542,6 +602,20 @@ const initialSessionState: SessionState = {
     showSelectedVariableLayer: true,
     savedLayers: {},
 
+    selectedDataStoreId: null,
+    selectedDataSourceId: null,
+    dataSourceFilterExpr: '',
+    selectedOperationName: null,
+    operationFilterTags: [],
+    operationFilterExpr: '',
+
+    showDataSourceDetails: true,
+    showResourceDetails: true,
+    showWorkflowStepDetails: true,
+    showOperationDetails: true,
+    showVariableDetails: true,
+    showLayerDetails: true,
+
     panelContainerUndockedMode: false,
     leftPanelContainerLayout: {horPos: 300, verPos: 400},
     rightPanelContainerLayout: {horPos: 300, verPos: 400},
@@ -555,12 +629,14 @@ const initialSessionState: SessionState = {
     },
     selectedPlacemarkId: null,
     showPlacemarkDetails: true,
+    placemarkCounter: 0,
 
     workspacePanelMode: 'steps',
 
     showDataSourceTitles: true,
     showLayerTextOverlay: true,
     debugWorldView: false,
+    vectorStyleMode: "layer",
 
     backendConfig: {
         dataStoresPath: null,
@@ -570,9 +646,21 @@ const initialSessionState: SessionState = {
 };
 
 
-let placemarkCounter = 0;
+const PLACEMARK_TITLE_PREFIX = "Placemark ";
+
 const sessionReducer = (state: SessionState = initialSessionState, action: Action) => {
     switch (action.type) {
+        case actions.SET_SELECTED_ENTITY_ID: {
+            const selectedEntityId = action.payload.selectedEntityId || null;
+            let selectedPlacemarkId = null;
+            if (isString(selectedEntityId) && selectedEntityId.startsWith('placemark-')) {
+                selectedPlacemarkId = selectedEntityId;
+            }
+            if (selectedPlacemarkId !== state.selectedPlacemarkId) {
+                return {...state, selectedPlacemarkId};
+            }
+            break;
+        }
         case actions.UPDATE_INITIAL_STATE:
             return {...state, ...action.payload.session};
         case actions.UPDATE_SESSION_STATE:
@@ -587,11 +675,15 @@ const sessionReducer = (state: SessionState = initialSessionState, action: Actio
         case actions.ADD_PLACEMARK: {
             const position = action.payload.position;
             const features = state.placemarkCollection.features.slice();
+            const placemarkCounter = state.placemarkCounter;
+            const letter = String.fromCharCode(65 + placemarkCounter % 26);
             const newPlacemark = {
                 type: 'Feature',
                 id: genPlacemarkId(),
                 properties: {
-                    name: String.fromCharCode(65 + placemarkCounter % 26),
+                    ...featurePropertiesFromSimpleStyle(PLACEMARKS_LAYER.style as any),
+                    "marker-symbol": letter,
+                    title: PLACEMARK_TITLE_PREFIX + letter,
                     visible: true,
                 },
                 geometry: {
@@ -599,10 +691,9 @@ const sessionReducer = (state: SessionState = initialSessionState, action: Actio
                     coordinates: position ? [position.longitude, position.latitude] : [0, 0]
                 }
             } as Placemark;
-            placemarkCounter++;
             features.push(newPlacemark);
             const placemarkCollection = {...state.placemarkCollection, features};
-            return {...state, placemarkCollection, selectedPlacemarkId: newPlacemark.id};
+            return {...state, placemarkCollection, placemarkCounter: placemarkCounter + 1, selectedPlacemarkId: newPlacemark.id};
         }
         case actions.REMOVE_PLACEMARK: {
             const placemarkId = action.payload.placemarkId;
@@ -622,18 +713,74 @@ const sessionReducer = (state: SessionState = initialSessionState, action: Actio
             }
             break;
         }
-        case actions.UPDATE_PLACEMARK: {
-            const placemark = action.payload.placemark;
-            const placemarkId = placemark.id;
-            let features = state.placemarkCollection.features;
+        case actions.UPDATE_PLACEMARK_GEOMETRY: {
+            const {placemarkId, geometry} = action.payload;
+            const features = state.placemarkCollection.features.slice();
             const featureIndex = features.findIndex(f => f.id === placemarkId);
+            const oldFeature = featureIndex >= 0 && features[featureIndex];
+            assert.ok(oldFeature);
+            const oldGeometry = oldFeature && oldFeature.geometry;
+            const newGeometry = {...oldGeometry, ...geometry};
+            const newFeature = {...oldFeature, geometry: newGeometry};
             if (featureIndex >= 0) {
-                features = features.slice();
-                const feature = features[featureIndex];
-                features[featureIndex] = {...feature, ...placemark};
-                const placemarkCollection = {...state.placemarkCollection, features};
-                return {...state, placemarkCollection};
+                features[featureIndex] = newFeature;
+            } else {
+                features.push(newFeature);
             }
+            const placemarkCollection = {...state.placemarkCollection, features};
+            return {...state, placemarkCollection};
+        }
+        case actions.UPDATE_PLACEMARK_PROPERTIES: {
+            const {placemarkId, properties} = action.payload;
+            const features = state.placemarkCollection.features.slice();
+            const featureIndex = features.findIndex(f => f.id === placemarkId);
+            const oldFeature = featureIndex >= 0 && features[featureIndex];
+            assert.ok(oldFeature);
+            const oldProperties = oldFeature && oldFeature.geometry;
+            const newProperties = {...oldProperties, ...properties};
+            const newFeature = {...oldFeature, properties: newProperties};
+            if (featureIndex >= 0) {
+                features[featureIndex] = newFeature;
+            } else {
+                features.push(newFeature);
+            }
+            const placemarkCollection = {...state.placemarkCollection, features};
+            return {...state, placemarkCollection};
+        }
+        case actions.UPDATE_PLACEMARK_STYLE: {
+            const {placemarkId, style} = action.payload;
+            const features = state.placemarkCollection.features.slice();
+            const featureIndex = features.findIndex(f => f.id === placemarkId);
+            const oldFeature = featureIndex >= 0 && features[featureIndex];
+            assert.ok(oldFeature);
+            const oldProperties = oldFeature && oldFeature.properties;
+            const newProperties = {...oldProperties, ...featurePropertiesFromSimpleStyle(style)};
+            const oldTitle = oldProperties && oldProperties["title"];
+            const newTitle = newProperties && newProperties["title"];
+            if (newTitle && newTitle !== oldTitle) {
+                // Update marker symbol as long as we don't have an editor for symbols
+                let text;
+                if (newTitle.startsWith(PLACEMARK_TITLE_PREFIX)) {
+                    text = newTitle.substr(PLACEMARK_TITLE_PREFIX.length);
+                } else {
+                    text = newTitle;
+                }
+                let markerSymbol = newProperties["marker-symbol"];
+                if (text.length > 0) {
+                    markerSymbol = text[0];
+                } else if (!markerSymbol) {
+                    markerSymbol = "?";
+                }
+                newProperties["marker-symbol"] = markerSymbol;
+            }
+            const newFeature = {...oldFeature, properties: newProperties};
+            if (featureIndex >= 0) {
+                features[featureIndex] = newFeature;
+            } else {
+                features.push(newFeature);
+            }
+            const placemarkCollection = {...state.placemarkCollection, features};
+            return {...state, placemarkCollection};
         }
 
     }
@@ -676,10 +823,10 @@ const initialLocationState: LocationState = {
 
 //noinspection JSUnusedLocalSymbols
 const locationReducer = (state: LocationState = initialLocationState, action: Action) => {
-    if (action.type === SET_GLOBE_MOUSE_POSITION) {
+    if (action.type === actions.SET_GLOBE_MOUSE_POSITION) {
         const globeMousePosition = action.payload.position;
         return {...state, globeMousePosition};
-    } else if (action.type === SET_GLOBE_VIEW_POSITION) {
+    } else if (action.type === actions.SET_GLOBE_VIEW_POSITION) {
         const globeViewPosition = action.payload.position;
         return {...state, globeViewPosition};
     }
@@ -691,9 +838,9 @@ const locationReducer = (state: LocationState = initialLocationState, action: Ac
 // Combined State reducer
 
 export const stateReducer = combineReducers<State>({
-    data: dataReducer as Reducer<DataState>,
-    control: controlReducer as Reducer<ControlState>,
-    session: sessionReducer as Reducer<SessionState>,
-    communication: communicationReducer as Reducer<CommunicationState>,
-    location: locationReducer as Reducer<LocationState>,
-});
+                                                       data: dataReducer as Reducer<DataState>,
+                                                       control: controlReducer as Reducer<ControlState>,
+                                                       session: sessionReducer as Reducer<SessionState>,
+                                                       communication: communicationReducer as Reducer<CommunicationState>,
+                                                       location: locationReducer as Reducer<LocationState>,
+                                                   });
